Created provenance when pipeline is finished
