+++ packages/coinstac-ui/app/main/index.js
@@ -5,14 +5,19 @@
  * kicking off the render process (UI).
  */
 
+/* eslint-disable no-await-in-loop */
+
 'use strict';
 
 Error.stackTraceLimit = 100;
 
-const { compact } = require('lodash'); // eslint-disable-line no-unused-vars
+const {
+  compact, keys, pick, omit,
+} = require('lodash'); // eslint-disable-line no-unused-vars
 const electron = require('electron');
 const ipcPromise = require('ipc-promise');
 const fs = require('fs');
+const fsPromises = require('fs').promises;
 const path = require('path');
 const moment = require('moment');
 const ipcFunctions = require('./utils/ipc-functions');
@@ -55,6 +60,37 @@ const getAllFilesInDirectory = (directory) => {
     return dirent.isDirectory() ? getAllFilesInDirectory(res) : res;
   });
   return Array.prototype.concat(...files);
+};
+
+const exists = async (path) => {
+  const result = await fsPromises.stat(path).catch(() => false);
+
+  return Boolean(result);
+};
+
+const generateRunProvenance = async (userId, appDirectory, run, consortium) => {
+  const clients = keys(run.clients).map(clientId => ({
+    id: clientId,
+    username: run.clients[clientId],
+  }));
+
+  const provenance = {
+    ...omit(run, ['clients', 'consortiumId', '__typename']),
+    consortium: pick(consortium, ['id', 'name']),
+    clients,
+  };
+
+  const runProvenanceDirectory = path.join(appDirectory, 'output', userId, provenance.id);
+
+  if (!await exists(runProvenanceDirectory)) {
+    await fsPromises.mkdir(runProvenanceDirectory, { recursive: true });
+  }
+
+  const provenanceFilePath = path.join(runProvenanceDirectory, 'provenance.json');
+
+  if (!await exists(provenanceFilePath)) {
+    await fsPromises.writeFile(provenanceFilePath, JSON.stringify(provenance, null, 2));
+  }
 };
 
 let initializedCore;
@@ -198,25 +234,6 @@ loadConfig()
       mainWindow.webContents.send('prepare-consortia-files', res);
     });
 
-    /**
-     * IPC Listener to generate provenance
-     */
-    ipcMain.on('generate-provenance', async (event, { userId, appDirectory, runs }) => {
-      runs.forEach((run) => {
-        const runProvenanceDirectory = path.join(appDirectory, 'output', userId, run.id);
-
-        if (!fs.existsSync(runProvenanceDirectory)) {
-          fs.mkdirSync(runProvenanceDirectory, { recursive: true });
-        }
-
-        const provenanceFilePath = path.join(runProvenanceDirectory, 'provenance.json');
-
-        if (!fs.existsSync(provenanceFilePath)) {
-          fs.writeFileSync(provenanceFilePath, JSON.stringify(run, null, 2));
-        }
-      });
-    });
-
     ipcPromise.on('login-init', ({
       userId, appDirectory, clientServerURL, token,
     }) => {
@@ -358,6 +375,10 @@ loadConfig()
                 ipcFunctions.sendNotification(
                   'Pipeline finished',
                   `Pipeline ${pipelineName} finished on consortia ${consortiumName}`
+                );
+
+                generateRunProvenance(
+                  initializedCore.clientId, initializedCore.appDirectory, run, consortium
                 );
 
                 return initializedCore.unlinkFiles(run.id)
